Update SECURITY.md
Changed a Code line. Developed a new simulator for Samsung and GitHub can make come to life.  I'm going to Trust the that my effects to work with you will fall in vain but I've never none to be wrong. So let's test will this be the first time I got something wrong . God's speed on this project. Part of it is classified when we get to the war when and where, name's and ranks can be made up.  Just add the ability to change everything as a root option. So gov won't have to depend on use for parts ect.  As well if this sales abroad which it will a area of caution has to apply because of reason of platforms accusing each other in actual reality. What we are about to achieve is something crazy but real as it can get . Because we will be making hardware that can decide weather one of are come home this talk about the product ends here. 
To end that you armed forces around the world for bringing humanity home . You have are support in your mission. Therefore we are responding to make training easier and advanced at the same time. 
Teams that want to work with me on this can contact me at 2066379370 leave a text message before you call . Or email me at gine412yiri@post.wordpress.com
I go by Thanos or Gordo.