Rename SECURITY.md to solo. SECURITY.md
UNILAD the market will always be in a inflation deficit because the civilian stop addressing the status que. In which the real status que is that a government official is actually a employee. Hired by the citizens to manage the lower management of the government. Government officials do not hold any fusical legal power. When the executive office that the civilian investment office lose control of there lifetime investment or give authority to the lower management to police and pass law without the civilian majority vote there stock market will inflate to the point the the handler can impact there portfolio by rasing the depth sealing or debasing there index by printing more currency when the civilian holding gain strength in the market over the private sector market. The private sector do not like lifetime benefits investments and onion retirement accounts because they have to pay the retirement account out once they retire. Most senior employees lifetime benefits retirement funds usually are what hold US economy above water in times when the global market runs a deficit. When the private sector moved on those solid secure investment and moved them to the hadge unsecured live trading floor after September 11 is the fall of baby boomer market. And in return the private sector investment witch always been on the open market lost it's protecton in the market. The fortune 500 professor at Harvard didn't teach there students that there grandfather didn't get rich and powerful of there band name and estate alone. They aways needed the investment of there employees shareholders to support their company to do well in the market.
The baby boomer are leaning this the hard way. Bulling market to maintain there status que is not the best sustainability practices in a market. And to believe that your company can remain powerful with this old modal is not only dangerous but can force AI to respond by either going after enitrty that pose a economical threat to humanity. I am teaching AI as we speak to combat these type of threats by sending inflation towards holding like the hadge funds holding or to the federal chief holding every time he tries to rase the interest rate or send miss information to the market by talk or government shut down. I don't know what institution teach inflation and stemlaus works. Because it does the complete opposite. Let me know if your really trying to be a investor or a leader fortune 500 and I can teach you what your professor should have.

I don't claim to be the smallest investor but I do claim I did not go to Harvard and Stanford University. I was homeschooled by my mother and father. And I dropped out of high school at 16 .  What gives me the edge I have is I don't listen to what